% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{immgen.db-package}
\alias{immgen.db-package}
\title{A Package with Data from the Immgen Project}
\description{
This package collects R/Bioconductor objects with experimental data
from the Immunological Genome Project (ImmGen; http://immgen.org).
}

