% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immgen}
\alias{immgen}
\title{Expression data fro immune cells.}
\format{An ExpressionSet with 22135 features and 778 samples.}
\source{
\url{http://www.immgen.org}
}
\usage{
immgen
}
\description{
An ExpressionSet object containing expression levels for 217 immune cells.
}
\keyword{datasets}

